+++ src/PlaywrightSharp.Abstractions/Browser.cs
@@ -1,0 +1,23 @@
+namespace PlaywrightSharp
+{
+    /// <summary>
+    /// Browser names.
+    /// </summary>
+    public enum Browser
+    {
+        /// <summary>
+        /// Chromium.
+        /// </summary>
+        Chromium,
+
+        /// <summary>
+        /// Firefox.
+        /// </summary>
+        Firefox,
+
+        /// <summary>
+        /// Webkit
+        /// </summary>
+        Webkit,
+    }
+}
