+++ src/PlaywrightSharp.Abstractions/BrowserFetcher.cs
