+++ src/PlaywrightSharp.Abstractions/BrowserFetcherOptions.cs
