+++ src/PlaywrightSharp.Abstractions/Helpers/BrowserHelper.cs
@@ -1,7 +1,6 @@
 using System;
-using System.Diagnostics;
 using System.Threading.Tasks;
-using PlaywrightSharp.Transport;
+using PlaywrightSharp.Server;
 
 namespace PlaywrightSharp.Helpers
 {
