+++ src/PlaywrightSharp.Abstractions/PlaywrightSharp.Abstractions.csproj
@@ -40,6 +40,7 @@
     </ItemGroup>
     <ItemGroup>
         <PackageReference Include="Esprima" Version="1.0.1270" />
+        <PackageReference Include="Microsoft.Extensions.Logging" Version="2.2.0" />
         <PackageReference Include="System.Text.Json" Version="4.7.0" />
     </ItemGroup>
     <ItemGroup>
@@ -49,5 +50,10 @@
     <ItemGroup>
         <EmbeddedResource Include="Resources\injectedSource.ts" />
         <EmbeddedResource Include="Resources\zsSelectorEngineSource.ts" />
+    </ItemGroup>
+    <ItemGroup>
+      <Reference Include="Microsoft.Extensions.Logging.Abstractions, Version=3.1.0.0, Culture=neutral, PublicKeyToken=adb9793829ddae60">
+        <HintPath>..\..\..\..\..\..\..\..\usr\local\share\dotnet\packs\Microsoft.AspNetCore.App.Ref\3.1.2\ref\netcoreapp3.1\Microsoft.Extensions.Logging.Abstractions.dll</HintPath>
+      </Reference>
     </ItemGroup>
 </Project>
