+++ src/PlaywrightSharp.Abstractions/Server/BrowserPaths.cs
@@ -1,0 +1,20 @@
+namespace PlaywrightSharp.Server
+{
+    internal class BrowserPaths
+    {
+        public static string GetBrowsersPath(string executablePath)
+        {
+            throw new System.NotImplementedException();
+        }
+
+        public static string GetBrowserDirectory(string browserPath, BrowserDescriptor browser)
+        {
+            throw new System.NotImplementedException();
+        }
+
+        public static string GetExecutablePath(string browserPath, BrowserDescriptor browser)
+        {
+            throw new System.NotImplementedException();
+        }
+    }
+}
